Changing OpenOpps CNAME
`openopps.digitalgov.gov` has moved to OPM `openopps.usajobs.gov`. They are working on getting a proper 301 redirect set up on their end in the coming weeks, but in the meantime we are going to CNAME their old URL to the new URL so that users are not reaching a dead-end page.